#138 changed winRate from integer type to double typegit add .!

+++ src/main/java/ch/uzh/ifi/hase/soprafs24/entity/User.java
@@ -62,7 +62,7 @@ public class User implements Serializable {
   @Column double winRate;
 
   public double getWinRate() {return winRate;}
-  public void setWinRate(int winRate) {this.winRate = winRate;}
+  public void setWinRate(double winRate) {this.winRate = winRate;}
 
   public int getLengthPR() {return lengthPR;}
   public void setLengthPR(int lengthPR) {this.lengthPR = lengthPR;}
